Background materials/03-Labs/Lab-02-RAG-Pipeline/starter/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
